Feat: Atomización de MonsterForm
